Rearranging applications in the repository
