+++ CMakeLists.txt
@@ -67,21 +67,24 @@ daq_add_library(
 )
 
 ##############################################################################
-# Test Apps
+# Applications
 
-daq_add_application(tpg_algorithms_emulator_WIB2 tpg_algorithms_emulator_WIB2.cxx LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs CLI11::CLI11)
-daq_add_application(tpg_algorithms_emulator_WIBEth tpg_algorithms_emulator_WIBEth.cxx LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs CLI11::CLI11)
+daq_add_application(wib2_tpg_algorithms_emulator wib2_tpg_algorithms_emulator.cxx LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs CLI11::CLI11)
+daq_add_application(wibeth_tpg_algorithms_emulator wibeth_tpg_algorithms_emulator.cxx LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs CLI11::CLI11)
+
+daq_add_application(wib2_binary_frame_modifier wib2_binary_frame_modifier.cxx LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs)
+daq_add_application(wibeth_binary_frame_modifier wibeth_binary_frame_modifier.cxx LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs)
+
+daq_add_application(wib2_binary_frame_reader wib2_binary_frame_reader.cxx LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs)
+daq_add_application(wibeth_binary_frame_reader wibeth_binary_frame_reader.cxx LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs)
+
 
 ##############################################################################
 # Test Apps
 
-daq_add_application(TDEFileCreator TDEFileCreator.cxx TEST LINK_LIBRARIES fdreadoutlibs)
-daq_add_application(WIB2TestBench WIB2TestBench.cxx TEST LINK_LIBRARIES fdreadoutlibs)
-daq_add_application(WIB2BinaryFrameReader WIB2BinaryFrameReader.cxx TEST LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs)
-daq_add_application(WIBEthBinaryFrameReader WIBEthBinaryFrameReader.cxx TEST LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs)
-daq_add_application(WIB2AddFakeHits WIB2AddFakeHits.cxx TEST LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs)
-daq_add_application(WIBEthAddFakeHits WIBEthAddFakeHits.cxx TEST LINK_LIBRARIES fdreadoutlibs hdf5libs::hdf5libs)
-
+daq_add_application(tde_file_creator tde_file_creator.cxx TEST LINK_LIBRARIES fdreadoutlibs)
+daq_add_application(wib2_test_bench wib2_test_bench.cxx TEST LINK_LIBRARIES fdreadoutlibs)
+daq_add_application(wibeth_test_bench wibeth_test_bench.cxx TEST LINK_LIBRARIES fdreadoutlibs)
 
 ##############################################################################
