+++ apps/wibeth_binary_frame_reader.cxx
@@ -1,14 +1,13 @@
 /**
- * @file WIB2AddFakeHits.cxx: quick and dirty solution to modify an input binary file and add 
- * fake hits at a predefined frame and channel numbers 
- * Usage: ./WIB2AddFakeHits /nfs/home/aabedabu/dunedaq-v3.1.1dunedaq-v3.1.1_pre_source_id/dev/test_files/output_0_3.out 
+ * @file wibeth_binary_frame_reader.cxx: reads a WIBEth frame file
+ *  and prints the ADC values 
  *
  * This is part of the DUNE DAQ Application Framework, copyright 2022.
  * Licensing/copyright details are in the COPYING file that you should have
  * received with this code.
  */
 
-#include "fddetdataformats/WIB2Frame.hpp"
+#include "fddetdataformats/WIBEthFrame.hpp"
 
 #include <iostream>
 #include <fstream>
@@ -42,7 +41,7 @@ using namespace dunedaq::daqdataformats;
 void
 print_usage()
 {
-  TLOG() << "Usage: WIB2AddFakeHits <input_file_name> <input_channel>";
+  TLOG() << "Usage: wibeth_binary_frame_reader <input_file_name> ";
 }
 
 
@@ -52,7 +51,7 @@ public:
 
     FrameFile(const char* filename)
         : m_file(filename, std::ifstream::binary),
-          m_buffer(new char[sizeof(dunedaq::fddetdataformats::WIB2Frame)])
+          m_buffer(new char[sizeof(dunedaq::fddetdataformats::WIBEthFrame)])
     {
         if(m_file.bad() || m_file.fail() || !m_file.is_open()){
             throw std::runtime_error(std::string("Bad file ")+std::string(filename));
@@ -64,15 +63,15 @@ public:
         if(m_length==0){
             throw std::runtime_error("Empty file");
         }
-        //if(m_length%sizeof(dunedaq::fddetdataformats::WIB2Frame)!=0){
+        //if(m_length%sizeof(dunedaq::fddetdataformats::WIBEthFrame)!=0){
         //    throw std::runtime_error("File does not contain an integer number of frames");
         //}
-        m_n_frames=50;//m_length/sizeof(dunedaq::fddetdataformats::WIB2Frame);
+        m_n_frames=m_length/sizeof(dunedaq::fddetdataformats::WIBEthFrame);
 
-	// Reinterpret the frame as WIB2Frame
+	// Reinterpret the frame as WIBEthFrame
 	
 	m_file.read(m_buffer, m_file.eof());
-        m_wib2_frame = reinterpret_cast<dunedaq::fddetdataformats::WIB2Frame*>(m_buffer);
+        m_wibeth_frame = reinterpret_cast<dunedaq::fddetdataformats::WIBEthFrame*>(m_buffer);
     }
 
     ~FrameFile()
@@ -86,19 +85,19 @@ public:
     // Number of frames in the file
     size_t num_frames() const { return m_n_frames; }
 
-    dunedaq::fddetdataformats::WIB2Frame* get_wib2_frame() const { return m_wib2_frame; }
+    dunedaq::fddetdataformats::WIBEthFrame* get_wibeth_frame() const { return m_wibeth_frame; }
 
-    dunedaq::fddetdataformats::WIB2Frame* frame(size_t i)
+    dunedaq::fddetdataformats::WIBEthFrame* frame(size_t i)
     {
         if(i>=num_frames()) return nullptr;
         // Seek to the right place in the file
-        m_file.seekg(i*sizeof(dunedaq::fddetdataformats::WIB2Frame));
+        m_file.seekg(i*sizeof(dunedaq::fddetdataformats::WIBEthFrame));
         // Check we didn't go past the end
         if(m_file.bad() || m_file.eof()) return nullptr;
         // Actually read the fragment into the buffer
-        m_file.read(m_buffer,sizeof(dunedaq::fddetdataformats::WIB2Frame));
+        m_file.read(m_buffer,sizeof(dunedaq::fddetdataformats::WIBEthFrame));
         if(m_file.bad() || m_file.eof()) return nullptr;
-        return reinterpret_cast<dunedaq::fddetdataformats::WIB2Frame*>(m_buffer);
+        return reinterpret_cast<dunedaq::fddetdataformats::WIBEthFrame*>(m_buffer);
     }
     
 
@@ -107,7 +106,7 @@ public:
 protected:
     std::ifstream m_file;
     char* m_buffer;
-    dunedaq::fddetdataformats::WIB2Frame* m_wib2_frame = nullptr;
+    dunedaq::fddetdataformats::WIBEthFrame* m_wibeth_frame = nullptr;
     size_t m_length;
     size_t m_n_frames;
 };
@@ -119,14 +118,12 @@ int main(int argc, char** argv)
 {
  
 
-  if (argc != 3) {
+  if (argc != 2) {
     print_usage();
     return 1;
   }
 
   const std::string ifile_name = std::string(argv[1]);
-
-  const size_t input_ch = atoi(argv[2]);
 
   // Read file
   FrameFile input_file = FrameFile(ifile_name.c_str()); 
@@ -134,40 +131,21 @@ int main(int argc, char** argv)
   std::cout << "Size of the input file " << input_file.length() << std::endl;
   std::cout << "Number of frames " << input_file.num_frames() << std::endl;
    
-  size_t frame_number = 0; // frame chosen randomly
-  size_t ch = input_ch;
 
-  
-  dunedaq::fddetdataformats::WIB2Frame* frame = input_file.frame(frame_number);
-  
-  auto val_to_check = frame->get_adc(ch);
-  std::cout << "Output adc value: " << val_to_check << " for frame " << frame_number << " and channel number " << ch << std::endl;
-
- 
-  // Write output file 
-  // Modified two channels in the input
-  std::fstream output_file;
-  output_file.open("modified_output.bin", std::ios::app | std::ios::binary);
-
-  dunedaq::fddetdataformats::WIB2Frame* output_frame; 
+  dunedaq::fddetdataformats::WIBEthFrame* output_frame; 
   for (size_t i=0; i<input_file.num_frames(); i++) {
+    std::cout << "========== FRAME_NUM " << i <<  std::endl;
+    output_frame = input_file.frame(i);
+    for (int itime=0; itime<64; ++itime) {
+      for (int ch=0; ch<64; ++ch) {
+        uint16_t adc_val = output_frame->get_adc(ch, itime);
+        std::cout << "Output ADC value: " << adc_val << "\t\t\tFrame: " << i << " \t\tChannel: " << ch << " \t\tTimeSample: " << itime <<  std::endl;
+      }
+   }
+  }
 
-     
-     output_frame = input_file.frame(i);
-     auto val_to_check2 = output_frame->get_adc(ch);
-     std::cout << "Output adc value: " << val_to_check2 << " for frame " << i << " and channel number " << ch << std::endl;
-
-
-     if (i==frame_number || i==frame_number+1 || i==frame_number+2) {
-       auto adc_val = output_frame->get_adc(242);
-       auto adc_val2 = output_frame->get_adc(12);
-       output_frame->set_adc(242, adc_val+500);
-       output_frame->set_adc(12, adc_val2+500);
-     } 
-     output_file.write(reinterpret_cast<char*>(output_frame), sizeof(dunedaq::fddetdataformats::WIB2Frame) );
   
-  } 
 
-
+  
 }
